Add wrapper-structure 'v256d' for type '__m256d'
With the Intel intrinsic type '__m256d' it is complicated to access
individual elements of a vector register. Thus it is not suitable to use
them within a template structure.
By building this structure one can now provide abstract operations like
'+, -, *, /' and easier usage in source code.
At this time it is not clear if there is a performance penalty by using
this structure.